Move Collision Logic from Level to Entity
