+++ include/entity.h
@@ -6,6 +6,18 @@
 #include "physics.h"
 #include "transform.h"
 
+// Tile Type
+#define EMPTY 0
+#define PLATFORM 1
+#define SOLID_OBSTACLE 2
+#define SWITCH 3
+#define COINS 4
+#define SPIKE 5
+#define LIFE 6
+#define EXIT_GATE 9
+
+extern int solid_tiles[];
+
 typedef struct
 {
     SDL_Color color;
@@ -25,5 +37,11 @@ void update_entity(Entity *entity, float delta_time, Entity *objects[], int obje
 void apply_entity_movement(Entity *entity, float delta_time, Entity *objects[], int object_count);
 
 void destroy_entity(Entity *entity);
+
+bool is_solid(Transform *transform);
+
+bool is_void(Transform *transform);
+
+bool is_exit(Transform *transform);
 
 #endif
